==> 6_Identifikácia_používateľov.ipynb <==
#### Zadanie:

Usporiadajte najprv záznamy podľa IP, User-Agent a Unixtime.

Identifikujte používateľov na základe IP adresy a poľa User-Agent. Vytvorte novú premennú s ID používateľa.


Odovzdajte zdrojový kód programu + logovací súbor doplnený o identifikáciu používateľov (v jednom zip).

Príkazy v Pythone vhodné pre riešenie tohto zadania:

`
import pandas as pd
pd.read_csv() - načítanie log súboru
sort_values() - usporiadanie
`

import pandas as pd

#opening the file we prepared last time with having the Unixtime

file = pd.read_csv("5_UNIXTIME.csv", encoding="utf-8",  escapechar=";", index_col=False)

#must be sorted by "Remote Host", "UNIXTIME", "User-Agent" either in ascending or descending order
file.sort_values(by=['Remote Host', 'User-Agent', 'UNIXTIME'],ascending=[True, True, True], inplace=True)

# creating new array where the User ID aka identified session will be stored
user_id = []

#creating empty string variables which will be compared with the data from existing dataframe
IP_address = "0.0.0.0"
agent = "Null"
id=0

#with zip we create an iterable touple form the dataframe using 2 columns of interest
# which is "Remote Host" and "User-Agent"
for ip,user in zip(file["Remote Host"],file["User-Agent"]):
    
    #condition: only and only if the "Remote Host" and "User-Agent" is the same
    # as in last iteration, we can say that it is same session. Otherwise it is different one
    # because different users can go from the same IP, as well as from user agent
    # we omitt posibilities that 2 users from same IP and same user agent are executring uninterrupted session
    if (ip == IP_address) and (user == agent):
        user_id.append(id)
    else:
        id +=1
        user_id.append(id)
    IP_address = ip
    agent = user

#inserting the user_id array we created into the existing dataframe on the fist position

file.insert(0, 'USER_ID', user_id)

# writing outputs to the new csv

file.to_csv("6_Identifikácia_používateľov_USER_ID.csv", encoding="utf-8",  escapechar=";", index=False)
