==> 7_Premenná_čas_strávený_na_stránke_Length.ipynb <==
#### Zadanie:
Vytvorte premennú Length na základe User_ID a 60 min. SequenceID.


Odovzdajte zdrojový kód programu + logovací súbor doplnený o identifikáciu používateľov (v jednom zip).

Príkazy v Pythone vhodné pre riešenie tohto zadania:

`
import pandas as pd
pd.read_csv() - načítanie log súboru
`

import pandas as pd

# open the csv file where we have pripared data from which are separated by USER ID

file = pd.read_csv("6_Identifikácia_používateľov_USER_ID.csv", encoding="utf-8",  escapechar=";", index_col=False)

# prepare 2 empty arrays where we will be storing the new data columns "Sequence ID" and "Length"

sequence_id = []
lengths = []

# prepare temp variables which we need in the for loop to compare current and previous values
# starting value of all of them is assigned to 0 to make sense in the loop bcs
# user_id value starts with 1
# unixtime starts with some value and the difference between first value of unixtime and our temp variable needs to be 
# bigger than 3600

# seq_id is the Sequence ID which will be assigned to session
# length is the temp variable where we will store the difference in current and previous unixtime
# if the differnce is less than 3600 between them
# otherwise we will write value for None in list lengths
previous_user_id = 0
previous_unixtime = 0
seq_id = 0
length = 0
counter = 0

for user_id, unixtime in zip(file["USER_ID"],file["UNIXTIME"]):
    if (user_id == previous_user_id) and ((unixtime - previous_unixtime) <= 3600):
        length = unixtime - previous_unixtime
        sequence_id.append(seq_id)
        lengths.insert((counter-1), length)
        previous_unixtime = unixtime
        previous_user_id = user_id
        counter += 1
    else: 
        seq_id += 1
        sequence_id.append(seq_id)
        previous_unixtime = unixtime
        previous_user_id = user_id
        lengths.insert((counter-1), None)
        counter += 1

# inserting the columns on locations where we want them
# SEQUENCE_ID will become second behing the USER_ID and Length will go after UNIXTIME

file.insert(1, 'SEQUENCE_ID', sequence_id)
file.insert(7, 'Length', lengths)

# writing the output to the file
file.to_csv("7_Premenná_čas_strávený_na_stránke_Length.csv", encoding="utf-8",  escapechar=";", index=False)
