==> 8_Identifikácia_sedení_RLength.ipynb <==
#### Zadanie:
Identifikujte sedenia na základe heuristiky RLength (podiel navigačných stránok 40%).


Odovzdajte zdrojový kód programu + logovací súbor doplnený o identifikáciu používateľov (v jednom zip).

Príkazy v Pythone vhodné pre riešenie tohto zadania:

`import pandas as pd
import numpy as np
pd.read_csv() - načítanie log súboru
np.log - funkcia na výpočet prirodzeného logaritmu (ln)`

import pandas as pd
import numpy as np

# opening the file we will be working with and assigning it to the variabel file for latter use
file = pd.read_csv("7_Premenná_čas_strávený_na_stránke_Length.csv",  encoding="utf-8",  escapechar=";", index_col=False)

# calculating the mean of the time spent on the webpages from the pages where value is different than None
# tried and it gives here the same result as np.nanmean, despite it sais that np.nanmean ignores the NaN, while mean keep them
mean = np.mean(file["Length"])

mean

# C = -ln(1-p) / lambda
# p = podiel navigačných stránok 40%
p = 0.4
C = (-(np.log(1-p)))/(1/mean)

# value for the border point (hranicny bod)
C

# initializing the empty array for RLenght id numbers
# initializing the temp var RL_id to store temp ID numbers of the RLenght position index
RLength = []
RL_id = 0

# temp variables for comparison
previous_seq_id = 0
previous_length = 0
# we increment when seq_id != previous_seq_id

# we increment when seq_id == previous_seq_id and:
#     1. previous_length is NaN and length variable is bigger than C value
#     2. previous_length is NaN and length is NaN
#     3. previous_length is more than C 
#     4. previous_length is less than C and length is NaN

for seq_id, length in zip(file["SEQUENCE_ID"], file["Length"]):
    if ((previous_seq_id != seq_id) 
        or ((previous_seq_id == seq_id) and (np.isnan(previous_length) and length > C))
        or ((previous_seq_id == seq_id) and (np.isnan(previous_length) and np.isnan(length)))
        or ((previous_seq_id == seq_id) and (previous_length > C))
        or ((previous_seq_id == seq_id) and (np.isnan(previous_length) and length < C))):     
        
        RL_id += 1
        RLength.append(RL_id)
        previous_seq_id = seq_id
        previous_length = length

    else:
        RLength.append(RL_id)
        previous_seq_id = seq_id
        previous_length = length

file.insert(8, 'RLength', RLength)

# write to file
file.to_csv("8_Identifikácia_sedení_RLength.csv", encoding="utf-8",  escapechar=";", index=False)
