==> 13_Dopĺňanie_ciest_STT_Mean.ipynb <==
#### Zadanie:

Doplňte cesty do očisteného logovacieho súboru, v ktorom boli identifikovaný používatelia a sedenia pomocou metódy STT_Mean.

Odovzdajte zdrojový kód programu + logovací súbor doplnený o identifikáciu používateľov (v jednom zip).

import pandas as pd
from urllib.parse import urlparse

file = pd.read_csv("12_Identifikácia_sedení_SLength.csv", escapechar = ";", encoding ="utf-8", index_col = False)

# I will use here www.example.com, but it referes to your domain name, which access logs are being examined
# (case specific)

# we need to create same structure of the data for the one we are searching 
# there are 2 ways. Either transform Request by adding https://www.example.com to each or delete
# the https://www.example.com from the referer. This way we will normalize them and can compare at latter instance

# one more rule is added., After trial and error I found that referer is one of 3 which can be treated as same:
# 1. https://www.example.com/
# 2. https://example.com
# 3. http://www.example.com

new_referer = []
for refer in file['Referer']:
    if (pd.isna(refer) or not refer.startswith("https://www.example.com/")
    or not refer.startswith("https://example.com/")
    or not refer.startswith("http://www.example.com/")):
        new_referer.append(refer)
    else:        
        referer_path = urlparse(refer).path + urlparse(refer).params + urlparse(refer).query + urlparse(refer).fragment
        new_referer.append(referer_path)


file['Referer'] = new_referer

df_STT_mean = file

rows = df_STT_mean.iloc[0:0]
STT_mean = 0
df_temp = df_STT_mean.iloc[0:0]

for index, row in df_STT_mean.iterrows():
    if(row.STT_mean == STT_mean):
        if(len(rows) and (rows['Request Uri'].iloc[-1] == row['Referer'])):
            rows = rows.append(row,ignore_index=True)
        elif((row['Request Uri'] in list(rows['Request Uri'])) and (not pd.isna(row['Referer'])) and str(row['Referer'].startswith('/'))):
            last_index = rows['Request Uri'].where(rows['Request Uri']==row['Request Uri']).last_valid_index()
            if (rows['Referer'].iloc[last_index] == row['Referer']):
                part_refer = list(rows['Request Uri'].iloc[last_index:])
                part_refer.reverse()
                part_url = list(rows['Referer'].iloc[last_index:])
                part_url.reverse()
                part_tuples = list(zip(part_url,part_refer))
                for item in part_tuples:
                    new_row = rows.iloc[-1]
                    new_row['Request Uri'] = item[0]
                    new_row['Referer'] = item[1]
                    new_row['UNIXTIME'] += 1
                    new_row.length = 1
                    df_temp = df_temp.append(new_row,ignore_index=True)
                rows = df_STT_mean.iloc[0:0]
            else:
                rows = df_STT_mean.iloc[0:0]
                try:
                    if str(row['Referer'].startswith('/')):
                        rows = rows.append(row,ignore_index=True)
                except:
                    continue
        else:
            rows = df_STT_mean.iloc[0:0]
            try:
                if str(row['Referer'].startswith('/')):
                    rows = rows.append(row,ignore_index=True)
            except:
                continue
    else:      
        STT_mean = row.STT_mean
        rows = df_STT_mean.iloc[0:0]
        try:
            if str(row['Referer'].startswith('/')):
                rows = rows.append(row,ignore_index=True)
        except:
            continue

#df_temp.cookie = 'new log'
df_STT_mean = df_STT_mean.append(df_temp,ignore_index=True)
df_STT_mean.sort_values(by=['Remote Host', 'User-Agent', 'UNIXTIME'],ascending=[True, True, True], inplace=True)
df_STT_mean

# writing the file to new csv, escape char ";" to not get User-Agent divided on export/write
file.to_csv("13_Dopĺňanie_ciest_STT_Mean.csv", encoding="utf-8",  escapechar=";", index=False)
